add home page

+++ Layout/profile.php
@@ -4,7 +4,7 @@ if (!isset($_SESSION['email'])) {
     header('location:login.php');
 }
 include '../includes/profileFunction.php';
-$user=getProfileInfo($_SESSION['id_user']);
+$user = getProfileInfo($_SESSION['id_user']);
 editProfile();
 ?>
 
@@ -46,38 +46,38 @@ editProfile();
             <main role="main" class="col-md-9 ml-sm-auto col-lg-10 px-4">
                 <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
                     <h1 class="h2">Profile</h1>
-                    <a class="btn btn-danger" href="../includes/profileFunction.php?Delete_User=<?php echo $user['id_user'] ?>">Delete</a>
+                    <a class="btn btn-danger" href="../includes/profileFunction.php?Delete_User=<?php echo $user['id_user'] ?>" onclick="confirmDelete(event)">Delete</a>
                 </div>
                 <!-- Profile content -->
-                    <div class="col-12 p-5">
-                        <h1 class="text-center">My Information</h1>
-                        <form action="profile.php" method="post">
-                            <input type="hidden" name="id_user" value="<?php echo $user['id_user'];?>">
-                            <input type="hidden" name="mpDB" value="<?php echo $user['mp'];?>" >
-                            <div class="mb-3">
-                                <label for="fname" class="form-label">First Name</label>
-                                <input name="fname" value="<?php echo $user['first_name'];?>" type="text" class="form-control" id="fname">
-                            </div>
-                            <div class="mb-3">
-                                <label for="lname" class="form-label">Last Name</label>
-                                <input name="lname" value="<?php echo $user['last_name'];?>" type="text" class="form-control" id="lname">
-                            </div>
-                            <div class="mb-3">
-                                <label for="email" class="form-label">Email address</label>
-                                <input name="email" value="<?php echo $user['email'];?>" type="email" class="form-control" id="email" aria-describedby="emailHelp">
-                                <div id="emailHelp" class="form-text">We'll never share your email with anyone else.</div>
-                            </div>
-                            <div class="mb-3">
-                                <label for="CurrentPassword" class="form-label">Current Password</label>
-                                <input name="CurrentPassword" type="password" class="form-control" id="CurrentPassword">
-                            </div>
-                            <div class="mb-3">
-                                <label for="NewPassword" class="form-label">New Password</label>
-                                <input name="NewPassword" type="password" class="form-control" id="NewPassword">
-                            </div>
-                            <button type="submit" class="btn btn-primary">Save</button>
-                        </form>
-                    </div>
+                <div class="col-12 p-5">
+                    <h1 class="text-center">My Information</h1>
+                    <form action="profile.php" method="post">
+                        <input type="hidden" name="id_user" value="<?php echo $user['id_user']; ?>">
+                        <input type="hidden" name="mpDB" value="<?php echo $user['mp']; ?>">
+                        <div class="mb-3">
+                            <label for="fname" class="form-label">First Name</label>
+                            <input name="fname" value="<?php echo $user['first_name']; ?>" type="text" class="form-control" id="fname">
+                        </div>
+                        <div class="mb-3">
+                            <label for="lname" class="form-label">Last Name</label>
+                            <input name="lname" value="<?php echo $user['last_name']; ?>" type="text" class="form-control" id="lname">
+                        </div>
+                        <div class="mb-3">
+                            <label for="email" class="form-label">Email address</label>
+                            <input name="email" value="<?php echo $user['email']; ?>" type="email" class="form-control" id="email" aria-describedby="emailHelp">
+                            <div id="emailHelp" class="form-text">We'll never share your email with anyone else.</div>
+                        </div>
+                        <div class="mb-3">
+                            <label for="CurrentPassword" class="form-label">Current Password</label>
+                            <input name="CurrentPassword" type="password" class="form-control" id="CurrentPassword">
+                        </div>
+                        <div class="mb-3">
+                            <label for="NewPassword" class="form-label">New Password</label>
+                            <input name="NewPassword" type="password" class="form-control" id="NewPassword">
+                        </div>
+                        <button type="submit" class="btn btn-primary">Save</button>
+                    </form>
+                </div>
             </main>
         </div>
     </div>
@@ -100,7 +100,27 @@ editProfile();
 
     <!-- Graphs -->
     <script src="https://cdn.jsdelivr.net/npm/chart.js@2.7.1/dist/Chart.min.js"></script>
+    <!-- alerts -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/limonte-sweetalert2/10.16.6/sweetalert2.all.min.js"></script>
+    <script>
+        function confirmDelete(event) {
+            event.preventDefault(); // Prevent the default link behavior
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#d33',
+                cancelButtonColor: '#28282B',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
 
+                if (result.isConfirmed) {
+                        window.location.href = event.target.href;
+                }
+            })
+        }
+    </script>
 </body>
 
 </html>
